+++ src/Adapter/AMQP/AMQP.php
@@ -17,10 +17,6 @@ class AMQP implements AdapterInterface
      * @var Connection
      */
     protected $connection;
-    /**
-     * @var Channel
-     */
-    protected $channelPrototype;
 
     /**
      * AMQP constructor.
@@ -33,10 +29,9 @@ class AMQP implements AdapterInterface
     public function __construct($connection, Channel $channelPrototype = null)
     {
         if (!$connection instanceof Connection) {
-            $connection = new Connection($connection);
+            $connection = new Connection($connection, $channelPrototype);
         }
         $this->registerConnection($connection);
-        $this->registerChannel($channelPrototype ?: new Channel());
     }
 
     /**
@@ -48,51 +43,10 @@ class AMQP implements AdapterInterface
     }
 
     /**
-     * @param Channel $channel
-     */
-    public function registerChannel(Channel $channel)
-    {
-        $this->channelPrototype = $channel;
-    }
-
-    /**
      * @return Connection
      */
     public function getConnection()
     {
         return $this->connection;
-    }
-
-    /**
-     * @param \AMQPChannel $resource
-     * @return Channel
-     * @throws \AMQPConnectionException
-     */
-    public function createChannel($resource = null)
-    {
-        $channel = clone $this->channelPrototype;
-
-        $channel->setConnection($this->getConnection());
-
-        if ($resource instanceof \AMQPChannel) {
-            $channel->setResource($resource);
-        } else {
-            if (!$this->getConnection()->isConnected()) {
-                $this->getConnection()->connect();
-            }
-            $channel->setResource($this->createChannelResource());
-        }
-
-        return $channel;
-    }
-
-    /**
-     * @return \AMQPChannel
-     * @throws \AMQPConnectionException
-     * @codeCoverageIgnore
-     */
-    protected function createChannelResource()
-    {
-        return new \AMQPChannel($this->connection->getResource());
     }
 }