+++ src/Adapter/PhpAmqpLib/PhpAmqpLib.php
@@ -2,7 +2,6 @@
 
 namespace AMQPAL\Adapter\PhpAmqpLib;
 
-use PhpAmqpLib\Channel\AMQPChannel as LibChannel;
 use PhpAmqpLib\Connection\AbstractConnection as LibConnection;
 use AMQPAL\Adapter\AdapterInterface;
 use AMQPAL\Adapter\Exception;
@@ -15,10 +14,6 @@ class PhpAmqpLib implements AdapterInterface
      * @var Connection
      */
     protected $connection;
-    /**
-     * @var Channel
-     */
-    protected $channelPrototype;
 
     /**
      * PhpAmqpLib constructor.
@@ -32,10 +27,9 @@ class PhpAmqpLib implements AdapterInterface
     public function __construct($connection, Channel $channelPrototype = null)
     {
         if (!$connection instanceof Connection) {
-            $connection = new Connection($connection);
+            $connection = new Connection($connection, $channelPrototype);
         }
         $this->registerConnection($connection);
-        $this->registerChannel($channelPrototype ?: new Channel());
     }
 
     /**
@@ -44,38 +38,6 @@ class PhpAmqpLib implements AdapterInterface
     public function registerConnection(Connection $connection)
     {
         $this->connection = $connection;
-    }
-
-    /**
-     * @param Channel $channel
-     */
-    public function registerChannel(Channel $channel)
-    {
-        $this->channelPrototype = $channel;
-    }
-
-    /**
-     * @param LibChannel $resource
-     * @return Channel
-     * @throws Exception\RuntimeException
-     * @throws Exception\InvalidArgumentException
-     */
-    public function createChannel($resource = null)
-    {
-        $channel = clone $this->channelPrototype;
-
-        $channel->setConnection($this->getConnection());
-
-        if ($resource instanceof LibChannel) {
-            $channel->setResource($resource);
-        } else {
-            if (!$this->getConnection()->isConnected()) {
-                $this->getConnection()->connect();
-            }
-            $channel->setResource($this->getConnection()->getResource()->channel());
-        }
-
-        return $channel;
     }
 
     /**